+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `DocsFromParent` to __all__ in satella.coding
 * added extra methods to `ListDeleter`, along with supporting reversed protocol
+* `DocsFromParent` will fill in the docstring for the class itself